refactor: relocate story tests to component directories
